+++ setup.cfg
@@ -1,0 +1,30 @@
+[metadata]
+long_description = file: README.md
+keywords = 
+    moon
+    crater
+    detector
+    hydra
+    lighting
+    torch
+license = MIT License
+classifiers =
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3.8
+
+[options]
+zip_safe = True
+include_package_data = True
+packages = find:
+install_requires =
+    pytorch_lightning==1.5.10
+    torchvision==0.11.1
+    torchmetrics==0.7.2
+    omegaconf==2.1.1
+    hydra_core==1.1.1
+    craterdata==0.1.3
+    python-dotenv==0.19.2
+    rich==11.2.0
+
+[console_scripts]
+deepmoon = deepmoon.__main__
