+++ torchmoon/torch/util.py
